third_party/WebKit/LayoutTests/external/wpt/gyroscope/Gyroscope.https-expected.txt

This is a testharness.js-based test.
FAIL Gyroscope: Test that 'onreading' is called and sensor reading is valid assert_equals: Expected reading event, but got error event instead expected "reading" but got "error"
FAIL Gyroscope: sensor reading is correct assert_equals: Expected reading event, but got error event instead expected "reading" but got "error"
FAIL Gyroscope: sensor timestamp is updated when time passes assert_equals: Expected reading event, but got error event instead expected "reading" but got "error"
FAIL Gyroscope: Test that sensor can be successfully created and its states are correct. assert_equals: Expected activate event, but got error event instead expected "activate" but got "error"
FAIL Gyroscope: sensor.start() returns undefined assert_equals: Expected activate event, but got error event instead expected "activate" but got "error"
FAIL Gyroscope: no exception is thrown when calling start() on already started sensor assert_equals: Expected activate event, but got error event instead expected "activate" but got "error"
FAIL Gyroscope: sensor.stop() returns undefined assert_equals: Expected activate event, but got error event instead expected "activate" but got "error"
FAIL Gyroscope: no exception is thrown when calling stop() on already stopped sensor assert_equals: Expected activate event, but got error event instead expected "activate" but got "error"
FAIL Gyroscope: Test that fresh reading is fetched on start() assert_equals: Expected reading event, but got error event instead expected "reading" but got "error"
FAIL Gyroscope: sensor readings can not be fired on the background tab assert_equals: Expected reading event, but got error event instead expected "reading" but got "error"
FAIL Gyroscope: frequency hint works promise_test: Unhandled rejection with value: object "[object SensorErrorEvent]"
FAIL Gyroscope: sensor receives suspend / resume notifications when  cross-origin subframe is focused assert_equals: Expected reading event, but got error event instead expected "reading" but got "error"
FAIL Gyroscope: throw 'NotSupportedError' for an unsupported sensor option assert_throws: function "() => { new sensorType({invalid: 1}) }" did not throw
FAIL Gyroscope: throw 'TypeError' if frequency is invalid assert_throws: when freq is undefined function "() => { new sensorType({frequency: freq}) }" did not throw
FAIL Gyroscope: sensor reading is correct when options.referenceFrame is 'screen' assert_equals: Expected reading event, but got error event instead expected "reading" but got "error"
FAIL Gyroscope: throw 'TypeError' if referenceFrame is not one of enumeration values assert_throws: when refFrame is undefined function "() => { new sensorType({referenceFrame: refFrame}) }" did not throw
Harness: the test ran to completion.
